refactor: move `intercept` logic into separate file
